+++ terraform/core/variables.tf
@@ -10,12 +10,7 @@ variable "helm_repository" {
   default = "../../helm/core" #"https://goafabric.github.io/example-konstruction-kit/helm/charts/example/spring"
 }
 
-data "external" "server_arch_data" {
-  program = ["sh", "-c", "if kubectl version --output=json | grep -q 'linux/arm64'; then echo '{\"server_arch\":\"-arm64v8\"}'; else echo '{\"server_arch\":\"\"}'; fi"]
-}
-
 locals {
-  server_arch = data.external.server_arch_data.result["server_arch"]
   oidc_enabled = strcontains(var.hostname, ".de")
 }