+++ xargocd/core/core.tf
@@ -23,7 +23,7 @@ resource "kubernetes_manifest" "core-application" {
             },
             {
               name  = "image.arch"
-              value = "-native${local.server_arch}"
+              value = "-native"
             },
             {
               name  = "maxReplicas"
@@ -45,11 +45,11 @@ resource "kubernetes_manifest" "core-application" {
             },
             {
               name  = "database.password"
-              value = random_password.postgresql_password.result
+              value = data.kubernetes_secret.postgresql_secret.data["password"]
             },
             {
               name  = "s3.password"
-              value = "minioadmin"
+              value = data.kubernetes_secret.s3_secret.data["password"]
             },
             {
               name  = "messageBroker.password"